// src/app/modelos/header.interface.ts
export class headerI{
    banner: string='';
    logoYoP: string='./assets/images/logo1.png';
    textoYoP: string='#YoProgramo';
    enlaceG: string='https://github.com/foly13';
    logoG: string='bi bi-github';
    enlaceL: string='https://www.linkedin.com/in/florencia-paola-casanova-3710221b5/?originalSubdomain=ar';
    logoL: string='bi bi-linkedin';
    enlaceI: string='https://www.instagram.com/flor.casan/';
    logoI: string='bi bi-instagram';
    enlaceW: string='https://api.whatsapp.com/send?phone=5493415463439&text=Hola!%20';
    logoW: string='bi bi-whatsapp';
    id: number = 0;
}